Simplify minimization methods
